month-1/week-2/rag-from-scratch/pyproject.toml
[project]
name = "rag-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.5",
    "openai>=2.29.0",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
